requirements.txt
Django>=5.0,<6
Pillow>=10.0
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=22.0
whitenoise>=6.6
requests>=2.31
beautifulsoup4>=4.12
celery[redis]>=5.3,<6
redis>=5.0
